requirements.txt
feedparser==6.0.8
html2text==2020.1.16
newspaper3k==0.2.8
pandas==1.3.4
pytz==2021.3
requests==2.26.0
